Refactor: Move main.go to backend/cmd directory
